Refactor tests to verify logic id
Before this, the tests were checking to see if the
properties existed in the template, but not actually
checking if they were associated with the correct
resource. This is a gap in the cdk template testing
framework. This updates the tests to find a specific
resource by its logical id, and then verify that the
expected properties are present.