Feat/add provider users profile endpoint (#18)
This adds a new endpoint to allow provider users that have logged in 
to see their applicant dashboard. When a provider user is created, 
the process that created them will be responsible for setting their 
provider id and their compact as custom attributes on their cognito
user identity. These custom attributes will be used to query the 
database for their personal information.

* Refactor cdk tests to verify logic id

Before this, the tests were checking to see if the
properties existed in the template, but not actually
checking if they were associated with the correct
resource. This is a gap in the cdk template testing
framework. This updates the tests to find a specific
resource by its logical id, and then verify that the
expected properties are present.

* Move v1 api model definitions into api model class

To keep our schema setup separated from other configuration,
this moves all of the API GW model definitions for the v1 api
into their own class. This improves readability of the cdk logic
and sets up common source for multiple endpoints to share the same
model as needed.