Feat/add provider users profile endpoint (#25)
This adds an endpoint for provider users to see their profile
information when looking at the applicant dashboard

Before this, the tests were checking to see if the
properties existed in the template, but not actually
checking if they were associated with the correct
resource. This is a gap in the cdk template testing
framework. This updates the tests to find a specific
resource by its logical id, and then verify that the
expected properties are present.

* Move v1 api model definitions into api model class

To keep our schema setup separated from other configuration,
this moves all of the API GW model definitions for the v1 api
into their own class. This improves readability of the cdk logic
and sets up common source for multiple endpoints to share the same
model as needed.

This model is responsible for initializing the v1 api models
for the root api, which we then want to set as protected
so other classes won't modify it. This is a valid use case
for protected access to work with cdk.

* Update PR template to include CDK test item

Whenever we make API configuration changes, we should include at least one or more
CDK tests to improve the long-term maintainability of the system. This updates
the PR template to serve as a checklist item that reviewers can check for.