GET purchase privilege options (#32)
* Define configuration steps for states and compacts

Before a state can be onboarded to the Compact Connect system,
there is certain information that must be provided for
that state. This commit adds documentation outlining the
onboarding process, and provides configuration file placeholders

* Use separate compact config table

We want to keep our provider information separated from our compact
configuration in terms of where it is stored. This adds a new table
specifically for storing the compact config information, and adds new
config uploader and purchase lambdas to reference it.


* Add custom resource CDK definition for uploading compact config data

This loads all the yaml configuration files into a json string, which a lambda
function (to be added in next commit) will process to upload needed records
into the compact config dynamo table. To avoid users committing their sandbox
env names to the project this updates the lambda to always deploy ALL config 
files to their sandbox environment.


* Add custom resource directory with lambda and tests

This adds the logic for actually uploading the config data into
the Dynamo table. This custom-resources directory can be repurposed for
any other future custom resources we may add in the future.


* fix user pool email dependency mapping

There were race conditions where the user pool would
be stood up before the email identity was finished
setting up its DNS records. This adds an explicit
dependency on the identity and DMARC record itself.